Move  methods into the  package
